# swagger_server/models/treasure.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Treasure(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, location: List[float]=None, hint: object=None, found: List[object]=None):  # noqa: E501
        """Treasure - a model defined in Swagger

        :param location: The location of this Treasure.  # noqa: E501
        :type location: List[float]
        :param hint: The hint of this Treasure.  # noqa: E501
        :type hint: object
        :param found: The found of this Treasure.  # noqa: E501
        :type found: List[object]
        """
        self.swagger_types = {
            'location': List[float],
            'hint': object,
            'found': List[object]
        }

        self.attribute_map = {
            'location': 'location',
            'hint': 'hint',
            'found': 'found'
        }

        self._location = location
        self._hint = hint
        self._found = found

    @classmethod
    def from_dict(cls, dikt) -> 'Treasure':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Treasure of this Treasure.  # noqa: E501
        :rtype: Treasure
        """
        return util.deserialize_model(dikt, cls)

    @property
    def location(self) -> List[float]:
        """Gets the location of this Treasure.


        :return: The location of this Treasure.
        :rtype: List[float]
        """
        return self._location

    @location.setter
    def location(self, location: List[float]):
        """Sets the location of this Treasure.


        :param location: The location of this Treasure.
        :type location: List[float]
        """

        self._location = location

    @property
    def hint(self) -> object:
        """Gets the hint of this Treasure.


        :return: The hint of this Treasure.
        :rtype: object
        """
        return self._hint

    @hint.setter
    def hint(self, hint: object):
        """Sets the hint of this Treasure.


        :param hint: The hint of this Treasure.
        :type hint: object
        """

        self._hint = hint

    @property
    def found(self) -> List[object]:
        """Gets the found of this Treasure.


        :return: The found of this Treasure.
        :rtype: List[object]
        """
        return self._found

    @found.setter
    def found(self, found: List[object]):
        """Sets the found of this Treasure.


        :param found: The found of this Treasure.
        :type found: List[object]
        """

        self._found = found